+++ drivers/unix/dir_access_unix.h
@@ -79,7 +79,7 @@ public:
 	virtual Error rename(String p_path, String p_new_path);
 	virtual Error remove(String p_path);
 
-	virtual size_t get_space_left();
+	virtual uint64_t get_space_left();
 
 	virtual String get_filesystem_type() const;
